+++ src/network/network_client.cpp
@@ -565,7 +565,7 @@ NetworkRecvStatus ClientNetworkGameSocketHandler::SendDesyncLog(const std::strin
 	for (size_t offset = 0; offset < log.size();) {
 		Packet *p = new Packet(PACKET_CLIENT_DESYNC_LOG, SHRT_MAX);
 		size_t size = std::min<size_t>(log.size() - offset, SHRT_MAX - 2 - p->Size());
-		p->Send_uint16(size);
+		p->Send_uint16((uint16)size);
 		p->Send_binary(log.data() + offset, size);
 		my_client->SendPacket(p);
