+++ src/saveload/extended_ver_sl.cpp
@@ -473,7 +473,7 @@ static void Save_SLXI()
 		if (_sl_xv_feature_versions[info->index] > 0) {
 			item_count++;
 			length += 6;
-			length += SlCalcObjLength(info, _xlsi_sub_chunk_desc);
+			length += (uint32)SlCalcObjLength(info, _xlsi_sub_chunk_desc);
 			if (info->save_proc) {
 				uint32 extra_data_length = info->save_proc(info, true);
 				if (extra_data_length) {
@@ -644,7 +644,7 @@ static void loadVL(const SlxiSubChunkInfo *info, uint32 length)
 
 static uint32 saveVL(const SlxiSubChunkInfo *info, bool dry_run)
 {
-	uint32 length = strlen(_openttd_revision);
+	size_t length = strlen(_openttd_revision);
 	if (!dry_run) MemoryDumper::GetCurrent()->CopyBytes(reinterpret_cast<const byte *>(_openttd_revision), length);
 	return length;
 }
