+++ src/station_cmd.cpp
@@ -4451,7 +4451,7 @@ uint MoveGoodsToStation(CargoID type, uint amount, SourceType source_type, Sourc
 		});
 
 		uint to_deliver = amount - moving;
-		uint step_size = CeilDivT<uint>(to_deliver, used_stations.size());
+		uint step_size = CeilDivT<uint>(to_deliver, (uint)used_stations.size());
 		for (uint i = 0; i < used_stations.size() && to_deliver > 0; i++) {
 			uint delivery = std::min<uint>(to_deliver, step_size);
 			used_stations[i].second += delivery;
