+++ src/tbtr_template_vehicle_func.cpp
@@ -97,7 +97,7 @@ void BuildTemplateGuiList(GUITemplateList *list, Scrollbar *vscroll, Owner oid, 
 	}
 
 	list->RebuildDone();
-	if (vscroll) vscroll->SetCount(list->size());
+	if (vscroll) vscroll->SetCount((uint)list->size());
 }
 
 Money CalculateOverallTemplateCost(const TemplateVehicle *tv)
